services/keyword-ingestion/requirements.txt
# Keyword Ingestion Service Requirements

fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aio-pika>=9.3.0
minio>=7.2.0
pandas>=2.1.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
aiosqlite>=0.19.0
